Use double everywhere instead of long double.

+++ pulsarDb/EphComputer.h
@@ -35,9 +35,9 @@ namespace pulsarDb {
 
       void cancelPdot(timeSystem::AbsoluteTime & ev_time) const;
 
-      long double calcPulsePhase(const timeSystem::AbsoluteTime & ev_time) const;
+      double calcPulsePhase(const timeSystem::AbsoluteTime & ev_time) const;
 
-      long double calcOrbitalPhase(const timeSystem::AbsoluteTime & ev_time) const;
+      double calcOrbitalPhase(const timeSystem::AbsoluteTime & ev_time) const;
 
       void modulateBinary(timeSystem::AbsoluteTime & emission_time) const;
